Move imports into main grammar, with a separate entry point for parsing JS imports
