+++ lib/compiler/passes/generate-js.js
@@ -8,7 +8,7 @@ const VERSION = require("../../version");
 const { stringEscape, regexpClassEscape } = require("../utils");
 const { SourceNode } = require("source-map-generator");
 const GrammarLocation = require("../../grammar-location");
-const { parse: parseImport } = require("./js-imports.js");
+const { parse } = require("../../parser.js");
 
 /**
  * @typedef {import("../../peg")} PEG
@@ -860,7 +860,13 @@ function generateJS(ast, options) {
               imports,
               code,
             ] = /** @type {PEG.ast.TopLevelInitializer[]} */ (
-              parseImport(tli.code)
+              parse(tli.code, {
+                startRule: "ImportsAndSource",
+                grammarSource: new GrammarLocation(
+                  tli.codeLocation.source,
+                  tli.codeLocation.start
+                ),
+              })
             );
             if (imports.code) {
               imps.push(imports);
